# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Storix API
  version: 0.1.0-draft
  description: |
    Storix는 호출 서버가 사용하는 독립 VFS(Virtual File System) 저장 서버다. 파일의
    업무적 의미와 최종 사용자 인증·권한 판단은 호출 서버의 책임이며 Storix 도메인에
    포함하지 않는다. 용어 정의는 `apps/api/CONTEXT.md` 참고.

    **초안 상태**: apps/demo로 이 API 모양이 실사용 검증되기 전에 작성됐다
    (`docs/ROADMAP.md`의 "실행 순서" 참고). demo 피드백에 따라 breaking change가
    발생할 수 있다.

    이 스펙은 `namespace`/`fs` 컨트롤러만 다룬다. `/health/*`(인증 예외 대상)와
    `/metrics`(운영 관측용)는 외부 API 계약이 아니라서 포함하지 않는다.

    **파일·스냅샷 경로 계약.** 공개 경로 입력은 `/`로 시작하는 절대경로다. TREE snapshot
    content의 `path`만 상대경로이며 `.`은 TREE 루트다. 중복 `/`, `.` 구간, 끝 `/`는
    제거해 같은 정규 경로로 식별한다. `..`, 백슬래시, C0·C1·DEL 제어 문자,
    Unicode Bidi_Control, 고립 surrogate 및 NFC가 아닌 이름은 정규화하지 않고
    400 `VFS_INVALID_PATH`로 거부한다. 이름은 대소문자를 구별하며 각각 UTF-8
    최대 255바이트다. 정규 절대경로는 앞 `/`를 포함해 UTF-8 최대 4096바이트이고,
    TREE 상대경로도 앞에 `/`를 붙인 길이로 검사한다. 이동·복사로 만들어질 최종
    경로와 모든 하위 경로에도 같은 한도를 적용한다. 경로 오류로 실패하면 파일
    트리·본문·snapshot은 변경되지 않는다. 조건부 변경의 오류 receipt는 해당
    endpoint의 재생 계약을 따른다. 부모 디렉터리는 기본적으로 생성하지 않으며
    `parents` 또는 `destinationParents`가 있는 작업에서 명시적으로 `true`로
    지정할 때만 생성한다. `/`는 조회·목록·TREE snapshot 원본과 이동·복사의
    디렉터리 목적지로 사용할 수 있다. 루트 자체를 파일로 쓰거나 이동·삭제할 수는 없다.

    **파일과 FILE snapshot 흐름.** 예를 들어 `/notebook.ipynb`의 새 바이트를
    `POST /api/v2/namespaces/{namespaceId}/fs/content/conditional?path=/notebook.ipynb`에
    `X-If-Absent: true`와 UUID 형식 `Idempotency-Key`·비어 있지 않은
    `X-Mutation-Scope`(UTF-8 최대 128바이트)로 보낸다.
...

...

...

    ```sh
...
    ```

...

...
paths:
...
  /api/v2/namespaces/{namespaceId}/fs/changes:
    get:
      operationId: listNamespaceChanges
      summary: namespace 변경 checkpoint 발급 및 변경 feed 페이지 조회
      description: |
        `change-feed` capability가 활성인 ACTIVE namespace에서만 조회한다. cursor 없이
        호출하면 현재 sequence의 checkpoint와 빈 changes를 반환한다. 이 checkpoint를
        전체 `ls` 열거 전에 보존하고, 열거 후 같은 cursor로 변경을 재생한다. cursor가
        있으면 해당 sequence 이후의 이벤트를 오름차순으로 반환한다. 각 페이지의
        `nextCursor`는 마지막 이벤트를 가리키고 빈 페이지에서는 입력 cursor를 유지한다.
        소비자는 페이지 적용과 cursor 저장을 원자적으로 처리하며 sequence로 중복을 제거한다.
        만료된 cursor는 버리고 checkpoint부터 전체 열거를 다시 한다.
      tags: [fs]
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
        - name: cursor
          in: query
          required: false
          description: 이전 응답의 불투명 nextCursor. 내부 형식을 해석하지 않는다.
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: checkpoint 또는 변경 페이지
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangeFeedPage' }
        '400':
          description: 잘못되거나 변조된 cursor 또는 다른 namespace cursor (`VFS_INVALID_CURSOR`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: change-feed capability 비활성 (`VFS_FEATURE_DISABLED`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '410':
          description: 보존 경계 이전 cursor (`VFS_CHANGE_CURSOR_EXPIRED`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalError' }
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        정적 API 키 슬라이스 인증(`apps/api/docs/adr/0007-service-auth-static-api-key.md`).
        배포 전체가 공유하는 단일 키 쌍(`STORIX_API_KEY`, 로테이션 중엔
        `STORIX_API_KEY_PREVIOUS`) 중 하나와 일치해야 한다. 실패 사유는
        구분해 알려주지 않고 항상 401만 반환한다.
  parameters:
...
    NamespaceId:
      name: namespaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      description: 페이지당 최대 항목 수. 유효하지 않은 값은 기본값으로 대체된다.
      schema:
        type: integer
        default: 100
        maximum: 1000
...
  responses:
...
    Unauthorized:
      description: '`Authorization` 헤더가 없거나 유효하지 않음(`UNAUTHORIZED`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Namespace 또는 경로를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ChangeFeedPage:
      type: object
      required: [changes, nextCursor, hasMore]
      properties:
        changes:
          type: array
          items: { $ref: '#/components/schemas/ChangeFeedEvent' }
        nextCursor: { type: string, description: 불투명 재개 cursor }
        hasMore: { type: boolean, description: 조회 시점의 다음 이벤트 존재 여부 }
    ChangeFeedEvent:
      type: object
      required: [sequence, operationId, operationIndex, operationCount, kind, nodeId, nodeType, path, occurredAt]
      properties:
        sequence: { type: string, pattern: '^[0-9]+$', description: namespace 내 십진 순서 번호 }
        operationId: { type: string, format: uuid }
        operationIndex: { type: integer, minimum: 0 }
        operationCount: { type: integer, minimum: 1 }
        kind: { type: string, enum: [created, updated, moved, deleted] }
        nodeId: { type: string, format: uuid }
        nodeType: { type: string, enum: [FILE, DIRECTORY] }
        path: { type: string, description: 삭제 이벤트에서는 삭제 직전 경로 }
        previousPath: { type: string, description: moved 이벤트에서만 존재 }
        revision: { type: string, description: 살아 있는 노드 이벤트에서만 존재 }
        occurredAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      description: |
        모든 4xx/5xx 응답의 공통 형태. RQ-018 범주의 공개 code, HTTP 상태, receipt·같은 key
        재시도 정책은 이 문서 상단의 “RQ-018 오류·재시도 계약” 표를 따른다. 그 밖의 endpoint 전용
        오류는 해당 operation 설명을 따른다. 500의 `message`는 고정된 안전 문구이며 내부 예외
        메시지나 비밀을 노출하지 않는다. `path`는 500 응답이나 요청 경로와 무관한 오류에서는
        생략될 수 있다.
      required: [code, message, requestId]
      properties:
        code:
          type: string
          example: VFS_NODE_NOT_FOUND
        message:
          type: string
        path:
          type: string
        requestId:
          type: string
          description: '`X-Request-Id` 요청 헤더 값 또는 서버가 생성한 값. 응답 `X-Request-Id` 헤더에도 echo된다.'
...